chore: clean up metrics
