+++ filecoin/log.go
@@ -1,0 +1,5 @@
+package filecoin
+
+import logging "github.com/ipfs/go-log/v2"
+
+var log = logging.Logger("filecoin")
