+++ filecoin/retriever.go
@@ -45,7 +45,6 @@ type RetrieverConfig struct {
 	MinerWhitelist     map[peer.ID]bool
 	DefaultMinerConfig MinerConfig
 	MinerConfigs       map[peer.ID]MinerConfig
-	Metrics            metrics.Metrics
 }
 
 func (cfg *RetrieverConfig) MinerConfig(peer peer.ID) MinerConfig {
@@ -101,11 +100,6 @@ func NewRetriever(
 	datastore datastore.Batching,
 	blockManager *blocks.Manager,
 ) (*Retriever, error) {
-
-	if config.Metrics == nil {
-		config.Metrics = metrics.NewNoop()
-	}
-
 	retriever := &Retriever{
 		config:                   config,
 		endpoint:                 endpoint,
@@ -128,19 +122,11 @@ func NewRetriever(
 // Possible errors: ErrInvalidEndpointURL, ErrEndpointRequestFailed,
 // ErrEndpointBodyInvalid, ErrNoCandidates
 func (retriever *Retriever) Request(cid cid.Cid) error {
-
-	requestInfo := metrics.RequestInfo{
-		RequestCid: cid,
-	}
-
 	// TODO: before looking up candidates from the endpoint, we could cache
 	// candidates and use that cached info. We only really have to look up an
 	// up-to-date candidate list from the endpoint if we need to begin a new
 	// retrieval.
 	candidates, err := retriever.lookupCandidates(context.Background(), cid)
-	retriever.config.Metrics.RecordGetCandidatesResult(requestInfo, metrics.GetCandidatesResult{
-		Err: err,
-	})
 	if err != nil {
 		return fmt.Errorf("could not get retrieval candidates for %s: %w", cid, err)
 	}
@@ -187,11 +173,6 @@ func (retriever *Retriever) retrieveFromBestCandidate(ctx context.Context, cid c
 	// complete
 	var retrievalStats *filclient.RetrievalStats
 	for _, query := range queries {
-		candidateInfo := metrics.CandidateInfo{
-			RequestInfo: metrics.RequestInfo{RequestCid: cid},
-			RootCid:     query.candidate.RootCid,
-			PeerID:      query.candidate.MinerPeer.ID,
-		}
 		if err := retriever.tryRegisterRunningRetrieval(query.candidate.RootCid, query.candidate.MinerPeer.ID); err != nil {
 			// TODO: send some info to metrics about this
 
@@ -202,8 +183,11 @@ func (retriever *Retriever) retrieveFromBestCandidate(ctx context.Context, cid c
 			continue
 		}
 
-		// TODO: Determine if we still need this log
-		retriever.config.Metrics.RecordRetrieval(candidateInfo)
+		log.Infof(
+			"Attempting retrieval from miner %s for %s",
+			query.candidate.MinerPeer.ID,
+			formatCidAndRoot(cid, query.candidate.RootCid, false),
+		)
 
 		stats.Record(ctx, metrics.RetrievalRequestCount.M(1))
 		stats.Record(ctx, metrics.RetrievalDealActiveCount.M(1))
@@ -213,21 +197,31 @@ func (retriever *Retriever) retrieveFromBestCandidate(ctx context.Context, cid c
 
 		stats.Record(ctx, metrics.RetrievalDealActiveCount.M(-1))
 		if err != nil {
+			log.Errorf(
+				"Failed to retrieve from miner %s for %s: %v",
+				query.candidate.MinerPeer.ID,
+				formatCidAndRoot(cid, query.candidate.RootCid, false),
+				err,
+			)
 			stats.Record(ctx, metrics.RetrievalDealFailCount.M(1))
 		} else {
+			log.Infof(
+				"Successfully retrieved from miner %s for %s\n"+
+					"\tDuration: %s\n"+
+					"\tBytes Received: %s\n"+
+					"\tTotal Payment: %s",
+				query.candidate.MinerPeer.ID,
+				formatCidAndRoot(cid, query.candidate.RootCid, false),
+				retrievalStats.Duration,
+				humanize.IBytes(retrievalStats.Size),
+				types.FIL(retrievalStats.TotalPayment),
+			)
+
 			stats.Record(ctx, metrics.RetrievalDealSuccessCount.M(1))
 			stats.Record(ctx, metrics.RetrievalDealDuration.M(retrievalStats.Duration.Seconds()))
 			stats.Record(ctx, metrics.RetrievalDealSize.M(int64(retrievalStats.Size)))
 			stats.Record(ctx, metrics.RetrievalDealCost.M(retrievalStats.TotalPayment.Int64()))
 		}
-
-		// TODO: Determine if we still need this log
-		retriever.config.Metrics.RecordRetrievalResult(candidateInfo, metrics.RetrievalResult{
-			Duration:      retrievalStats.Duration,
-			BytesReceived: retrievalStats.Size,
-			TotalPayment:  types.FIL(retrievalStats.TotalPayment),
-			Err:           err,
-		})
 
 		retriever.unregisterRunningRetrieval(query.candidate.RootCid, query.candidate.MinerPeer.ID)
 
@@ -400,18 +394,14 @@ func (retriever *Retriever) queryCandidates(ctx context.Context, cid cid.Cid, ca
 		go func(i int, candidate RetrievalCandidate) {
 			defer wg.Done()
 
-			candidateInfo := metrics.CandidateInfo{
-				RequestInfo: metrics.RequestInfo{RequestCid: cid},
-				RootCid:     candidate.RootCid,
-				PeerID:      candidate.MinerPeer.ID,
-			}
-
-			retriever.config.Metrics.RecordQuery(candidateInfo)
 			query, err := retriever.filClient.RetrievalQueryToPeer(ctx, candidate.MinerPeer, candidate.RootCid)
-			retriever.config.Metrics.RecordQueryResult(candidateInfo, metrics.QueryResult{
-				Err: err,
-			})
 			if err != nil {
+				log.Errorf(
+					"Failed to query miner %s for %s: %v",
+					candidate.MinerPeer.ID,
+					formatCidAndRoot(cid, candidate.RootCid, false),
+					err,
+				)
 				return
 			}
 
@@ -432,4 +422,21 @@ func (retriever *Retriever) queryCandidates(ctx context.Context, cid cid.Cid, ca
 
 func totalCost(qres *retrievalmarket.QueryResponse) big.Int {
 	return big.Add(big.Mul(qres.MinPricePerByte, big.NewIntUnsigned(qres.Size)), qres.UnsealPrice)
+}
+
+func formatCidAndRoot(cid cid.Cid, root cid.Cid, short bool) string {
+	if cid.Equals(root) {
+		return formatCid(cid, short)
+	} else {
+		return fmt.Sprintf("%s (root %s)", formatCid(cid, short), formatCid(root, short))
+	}
+}
+
+func formatCid(cid cid.Cid, short bool) string {
+	str := cid.String()
+	if short {
+		return "..." + str[len(str)-10:]
+	} else {
+		return str
+	}
 }
